Refactore drug order
